refactor: move power engines into infx.results.power
将单节点及多节点功耗引擎迁入 infx.results.power，合并共享的积分和产物处理逻辑，并保留旧命令及导入入口。
补充兼容性、精度及原子写入失败测试，并同步中英文架构文档。
