+++ infx/results/power/__init__.py
